Set `COMPOSE_PROJECT_NAME` in an isolated file.
What
=
Set `COMPOSE_PROJECT_NAME` in an isolated file, that gets overwritten
each time in full and is loaded by `~/.profile`.

Why
=
This makes the text being added to `~/.profile` consistent and unchanged
regardless of what the `COMPOSE_PROJECT_NAME` is set to, which means we
can reliably check to see if we've already made the changes we need to
to `~/.profile` even if the project name has changed.

Thanks
=
@kazgurs for the PR that led to this, #11.