What=Download and install the docker-compose version specified if it changes.How===Download docker-compose to a version path containing the version number,and symlink the version to a common path. Create the symlink on eachrun. If the compose version changes and that versions file is notpresent, the version will be downloaded.Why===It should be simple for the user to change the compose version withoutneeding to rebuilding the vagrant machine and without manually removingdocker-compose.In addition if a user switches back and forth between versions they shouldn't needto redownload the executable each time.
What
=
Download and install the docker-compose version specified if it changes.

How
===
Download docker-compose to a version path containing the version number,
and symlink the version to a common path. Create the symlink on each
run. If the compose version changes and that versions file is not
present, the version will be downloaded.

Why
===
It should be simple for the user to change the compose version without
needing to rebuilding the vagrant machine and without manually removing
docker-compose.

In addition if a user switches back and forth between versions they shouldn't need
to redownload the executable each time.