replace curl with net http (#55)
Replace usage and dependence on `curl` inside the container with having the plugin download the `docker-compose` binary on the host side, upload it to the guest machine.

As reported in #54 the plugin requires `curl` to be installed in the container. Not all images have `curl` installed and it's a little disappointing to require it to be installed in any image when we can download it in the plugin.