Enable binary mode writing the downloaded docker-compose binary to disk (#59)
### What

Enable binary mode writing the downloaded `docker-compose` binary to disk.

### Why

`File.open` accepts a `b` mode that indicates if file being written is a binary file. If the `b` mode isn't set the file will be written in text mode. Depending on the operating system writing and reading in text mode enables newline and encoding conversion. On Unix/Linux systems this has no impact, but on Windows systems the newline and encoding conversion kicks in. When writing the `docker-compose` binary on a a Windows system to the temporary file results in the binary being corrupted, because the file write is attempting to convert what it thinks is line breaks to Windows style line breaks.

Fixes #57